修改日志打印
