+++ shanghai/stock/stockheat/scala/com/kunyan/wokongsvc/realtimedata/SparkVisit.scala
@@ -12,7 +12,6 @@ import java.util.Calendar
 
 import com.kunyan.wokongsvc.realtimedata.CustomAccum._
 import com.kunyan.wokongsvc.realtimedata.JsonHandle._
-import com.kunyan.wokongsvc.realtimedata.logger.HeatLogger
 import kafka.serializer.StringDecoder
 import org.apache.log4j.PropertyConfigurator
 import org.apache.spark.rdd.RDD
@@ -30,12 +29,13 @@ object SparkVisit {
 
   def main(args: Array[String]) {
 
+    logger.info("开始执行")
     if (args.length != 2) {
-      HeatLogger.error("args too little")
+      logger.error("args too little")
       System.exit(-1)
     }
 
-    HeatLogger.warn("start execute")
+    logger.warn("start execute")
 
     /* 加载日志配置文件 */
     PropertyConfigurator.configure(args(0))
@@ -126,12 +126,12 @@ object SparkVisit {
 
               stockHandle.batchExec recover {
                 case e: Exception =>
-                  HeatLogger.exception(e)
+                 exception(e)
               }
 
               stockHandle.close()
 
-            case None => HeatLogger.warn("Get connect exception")
+            case None => logger.warn("Get connect exception")
           }
 
           otherStockPoolBr.value.getConnect match {
@@ -144,12 +144,12 @@ object SparkVisit {
 
               otherStockHandle.batchExec recover {
                 case e: Exception =>
-                  HeatLogger.exception(e)
+                 exception(e)
               }
 
               otherStockHandle.close()
 
-            case None => HeatLogger.warn("Get connect exception")
+            case None => logger.warn("Get connect exception")
           }
 
           testPoolBr.value.getConnect match {
@@ -162,12 +162,12 @@ object SparkVisit {
 
               testHandle.batchExec recover {
                 case e: Exception =>
-                  HeatLogger.exception(e)
+                 exception(e)
               }
 
               testHandle.close()
 
-            case None => HeatLogger.warn("Get connect exception")
+            case None => logger.warn("Get connect exception")
           }
 
         })
@@ -186,12 +186,12 @@ object SparkVisit {
 
                 mysqlHandle.batchExec recover {
                   case e: Exception =>
-                    HeatLogger.exception(e)
+                   exception(e)
                 }
 
                 mysqlHandle.close()
 
-              case None => HeatLogger.warn("Get connect exception")
+              case None => logger.warn("Get connect exception")
             }
           })
 
@@ -209,12 +209,12 @@ object SparkVisit {
 
                 mysqlHandle.batchExec recover {
                   case e: Exception =>
-                    HeatLogger.exception(e)
+                   exception(e)
                 }
 
                 mysqlHandle.close()
 
-              case None => HeatLogger.warn("Get connect exception")
+              case None => logger.warn("Get connect exception")
             }
           })
         }
