+++ shanghai/stock/stockheat/scala/com/kunyan/wokongsvc/realtimedata/Stock.scala
@@ -1,7 +1,6 @@
 package com.kunyan.wokongsvc.realtimedata
 
 import com.kunyan.wokongsvc.realtimedata.MixTool._
-import com.kunyan.wokongsvc.realtimedata.logger.HeatLogger
 
 import scala.collection.mutable
 import scala.util.{Failure, Success}
@@ -27,7 +26,7 @@ object Stock {
         val alias = sqlHandle.execQueryStockAlias(MixTool.SYN_SQL) match {
           case Success(z) => z
           case Failure(e) =>
-            HeatLogger.exception(e)
+            exception(e)
             System.exit(-1)
         }
         sqlHandle.close()
@@ -36,7 +35,7 @@ object Stock {
         alias
 
       case None =>
-        HeatLogger.error("[Get mysql connect failure]")
+        logger.error("[Get mysql connect failure]")
         System.exit(-1)
     }).asInstanceOf[Tuple2Map]
   }
@@ -50,7 +49,7 @@ object Stock {
         val stock = sqlHandle.execQueryStock(MixTool.STOCK_SQL) match {
           case Success(z) => z
           case Failure(e) =>
-            HeatLogger.exception(e)
+            exception(e)
             System.exit(-1)
         }
 
@@ -59,7 +58,7 @@ object Stock {
         stock
 
       case None =>
-        HeatLogger.error("[Get mysql connect failure]")
+        logger.error("[Get mysql connect failure]")
         System.exit(-1)
     }).asInstanceOf[mutable.HashSet[String]]
   }
